CURD_Operation/requirements.txt

python-decouple == 3.8
djangorestframework == 3.15.2
django-cors-headers == 4.6.0